Adds getParams function, Replaces url-matcher dep
