+++ examples/jaris-kitchen-sink/controllers/user.controller.ts
@@ -1,6 +1,10 @@
-import { Conn, json, status } from '@jaris/core';
-import { pipe } from '@jaris/util';
-import validate, { string, required, optional } from '@jaris/validator';
+import { Conn, json, status } from '../../../packages/core';
+import { pipe } from '../../../packages/util/src';
+import validate, {
+  string,
+  required,
+  optional,
+} from '../../../packages/validator/src';
 
 interface UserParams {
   userUid: string;
@@ -14,7 +18,6 @@ const createAndStoreUser = (userUid: string, body: any) => {
 const userController = {
   index: (conn: Conn) => {
     // ... fetch users
-    console.log('index');
     return json({ users: [] }, conn);
   },
   show: (conn: Conn, { userUid }: UserParams) => {
