+++ packages/router/package.json
@@ -18,8 +18,7 @@
   },
   "dependencies": {
     "@jaris/core": "^0.0.5",
-    "@jaris/util": "^0.0.5",
-    "url-matcher": "^0.2.2"
+    "@jaris/util": "^0.0.5"
   },
   "gitHead": "09a4395923c55d4d5bd55d35cc2aff60872c734f",
   "publishConfig": {
